Shrink two arrays

+++ common/dct.h
@@ -88,8 +88,8 @@ static const uint16_t x264_dct8_weight2_tab[64] = {
 };
 #undef W
 
-extern int x264_dct4_weight2_zigzag[2][16]; // [2] = {frame, field}
-extern int x264_dct8_weight2_zigzag[2][64];
+extern uint16_t x264_dct4_weight2_zigzag[2][16]; // [2] = {frame, field}
+extern uint16_t x264_dct8_weight2_zigzag[2][64];
 
 typedef struct
 {
